+++ workers/unity/Packages/io.improbable.gdk.transformsynchronization/MonoBehaviours/TransformSynchronization.cs
@@ -138,9 +138,9 @@ namespace Improbable.Gdk.TransformSynchronization
 
             var defaultToSet = new TransformToSet
             {
-                Position = TransformUtils.ToUnityVector3(transformComponent.Location) + world.GetExistingSystem<WorkerSystem>().Origin,
-                Velocity = TransformUtils.ToUnityVector3(transformComponent.Velocity),
-                Orientation = TransformUtils.ToUnityQuaternion(transformComponent.Rotation),
+                Position = transformComponent.Location.ToUnityVector() + world.GetExistingSystem<WorkerSystem>().Origin,
+                Velocity = transformComponent.Velocity.ToUnityVector(),
+                Orientation = transformComponent.Rotation.ToUnityQuaternion(),
                 ApproximateRemoteTick = 0
             };
 
@@ -162,9 +162,9 @@ namespace Improbable.Gdk.TransformSynchronization
 
             var defaultToSend = new TransformToSend
             {
-                Position = TransformUtils.ToUnityVector3(transformComponent.Location) - world.GetExistingSystem<WorkerSystem>().Origin,
-                Velocity = TransformUtils.ToUnityVector3(transformComponent.Velocity),
-                Orientation = TransformUtils.ToUnityQuaternion(transformComponent.Rotation)
+                Position = transformComponent.Location.ToUnityVector() - world.GetExistingSystem<WorkerSystem>().Origin,
+                Velocity = transformComponent.Velocity.ToUnityVector(),
+                Orientation = transformComponent.Rotation.ToUnityQuaternion()
             };
 
             var ticksSinceLastUpdate = new TicksSinceLastTransformUpdate