+++ workers/unity/Packages/io.improbable.gdk.transformsynchronization/Systems/RateLimitedTransformSendSystem.cs
@@ -47,22 +47,22 @@ namespace Improbable.Gdk.TransformSynchronization
 
                     var transformHasChanged = false;
 
-                    var newVelocity = TransformUtils.ToFixedPointVector3(transformToSend.Velocity);
-                    if (TransformUtils.HasChanged(newVelocity, transform.Velocity))
+                    var newVelocity = transformToSend.Velocity.ToFixedPointVector3();
+                    if (newVelocity != transform.Velocity)
                     {
                         transform.Velocity = newVelocity;
                         transformHasChanged = true;
                     }
 
-                    var newLocation = TransformUtils.ToFixedPointVector3(transformToSend.Position - worker.Origin);
-                    if (TransformUtils.HasChanged(newLocation, transform.Location))
+                    var newLocation = (transformToSend.Position - worker.Origin).ToFixedPointVector3();
+                    if (newLocation != transform.Location)
                     {
                         transform.Location = newLocation;
                         transformHasChanged = true;
                     }
 
-                    var newRotation = TransformUtils.ToCompressedQuaternion(transformToSend.Orientation);
-                    if (TransformUtils.HasChanged(newRotation, transform.Rotation))
+                    var newRotation = transformToSend.Orientation.ToCompressedQuaternion();
+                    if (newRotation != transform.Rotation)
                     {
                         transform.Rotation = newRotation;
                         transformHasChanged = true;